A number of small styling and usability fixes
